+++ src/main/java/com/jovan/erp_v1/request/WorkCenterRequest.java
@@ -1,12 +1,17 @@
 package com.jovan.erp_v1.request;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
 public record WorkCenterRequest(
-        Long id,
-        @NotBlank String name,
-        @NotBlank String location,
-        @NotNull @Positive Integer capacity) {
+                Long id,
+
+                @NotBlank String name,
+
+                @NotBlank String location,
+
+                @NotNull @Positive BigDecimal capacity) {
 }
