+++ src/main/java/com/jovan/erp_v1/service/IWorkCenterService.java
@@ -1,5 +1,6 @@
 package com.jovan.erp_v1.service;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.jovan.erp_v1.request.WorkCenterRequest;
@@ -19,21 +20,21 @@ public interface IWorkCenterService {
 
     List<WorkCenterResponse> findByName(String name);
 
-    List<WorkCenterResponse> findByCapacity(Integer capacity);
+    List<WorkCenterResponse> findByCapacity(BigDecimal capacity);
 
     List<WorkCenterResponse> findByLocation(String location);
 
     List<WorkCenterResponse> findByNameAndLocation(String name, String location);
 
-    List<WorkCenterResponse> findByCapacityGreaterThan(Integer capacity);
+    List<WorkCenterResponse> findByCapacityGreaterThan(BigDecimal capacity);
 
-    List<WorkCenterResponse> findByCapacityLessThan(Integer capacity);
+    List<WorkCenterResponse> findByCapacityLessThan(BigDecimal capacity);
 
     List<WorkCenterResponse> findByNameContainingIgnoreCase(String name);
 
     List<WorkCenterResponse> findByLocationContainingIgnoreCase(String location);
 
-    List<WorkCenterResponse> findByCapacityBetween(Integer min, Integer max);
+    List<WorkCenterResponse> findByCapacityBetween(BigDecimal min, BigDecimal max);
 
     List<WorkCenterResponse> findByLocationOrderByCapacityDesc(String location);
 }
